refactor(skill): split condition, unit timer, and readdb from skill.c
Move skill_check_condition2_pc, skill_unit_onplace_timer, and skill_readdb into separate files (Issue #56 phase 3). No behavior change.
