+++ vcproj/map-server.vcxproj
@@ -210,6 +210,9 @@
     <ClCompile Include="..\src\map\skill_castend_damage.c" />
     <ClCompile Include="..\src\map\skill_castend_ground.c" />
     <ClCompile Include="..\src\map\skill_castend_support.c" />
+    <ClCompile Include="..\src\map\skill_check_condition.c" />
+    <ClCompile Include="..\src\map\skill_readdb.c" />
+    <ClCompile Include="..\src\map\skill_unit_onplace.c" />
     <ClCompile Include="..\src\map\status.c" />
     <ClCompile Include="..\src\map\storage.c" />
     <ClCompile Include="..\src\map\trade.c" />
